Allows for the changing of an existing fw rule
netsh advfirewall firewall add doesn't have the capability to specify a
Grouping value; this firewall.ps1 allows us to use powershell to shuffle
around some of the rules from group to group.

Note: the rule will apply to ALL rules with matching names! Beware.
Also, there is not a way (yet) to clear the group value from an existing
rule.